add carousel to apartmentcomplexlayout
